refactor: test for login redirect
