+++ src/pages/auth/login/index.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
 
-import authorize from '../../../services/spotify/auth/authorize';
+import { loginRedirect } from './utils/loginRedirect';
 
 export function Login() {
   const location = useLocation<Record<string, unknown>>();
 
-  useEffect(() => authorize(location.state), [location]);
+  useEffect(() => loginRedirect(location.state), [location]);
   return <div data-testid="login-empty" />;
 }
