+++ src/pages/auth/login/utils/__tests__/verifyAuthCallback.test.ts
@@ -1,26 +1,22 @@
 import { URL, URLSearchParams } from 'url';
 
-import AuthenticationCallbackError from '../../../../errors/AuthenticationCallbackError';
-import getCurrentTimestamp from '../../../../utils/getCurrentTimestamp';
-import Storage from '../../../storage';
+import AuthenticationCallbackError from '../../../../../errors/AuthenticationCallbackError';
+import getCurrentTimestamp from '../../../../../utils/getCurrentTimestamp';
 import verifyAuthCallback from '../verifyAuthCallback';
 
-jest.mock('../../../../utils/getCurrentTimestamp');
-jest.mock('../../../storage');
+jest.mock('../../../../../utils/getCurrentTimestamp');
 
 const getCurrentTimestampMock = getCurrentTimestamp as jest.Mock;
 
 it('Read previous state from cookies', () => {
   getCurrentTimestampMock.mockReturnValue(3600);
-  const getItemSpy = jest
-    .spyOn(Storage, 'getItem')
-    .mockReturnValue('{"foo": "bar"}');
+  window.localStorage.setItem('randomIdHere', JSON.stringify({ foo: 'bar' }));
   const authorizeUrl = new URL('http://localhost:3000/auth/callback');
   const hashParams = new URLSearchParams();
   hashParams.append('access_token', 'foobar');
   hashParams.append('token_type', 'Bearer');
   hashParams.append('expires_in', '3600');
-  hashParams.append('state', 'randomThingHere');
+  hashParams.append('state', 'randomIdHere');
   hashParams.append('xyz', 'www');
   authorizeUrl.hash = hashParams.toString();
   const params = verifyAuthCallback(authorizeUrl.toString());
@@ -29,7 +25,6 @@ it('Read previous state from cookies', () => {
     state: { foo: 'bar' },
     token: 'foobar',
   });
-  getItemSpy.mockRestore();
 });
 
 it('Should return valid callback params from url', () => {
