+++ src/pages/auth/login/utils/openidConfiguration.ts
