--- src/services/storage/__mocks__/index.ts
@@ -1,3 +1,0 @@
-export default {
-  getItem: jest.fn(),
-};
