--- src/services/storage/index.ts
@@ -1,1 +1,0 @@
-export default window.localStorage;
